+++ src/main/java/org/gridsuite/filter/wip/rule/StringExpertRule.java
@@ -6,54 +6,60 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package org.gridsuite.filter.wip.expert.rule;
+package org.gridsuite.filter.wip.rule;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.annotations.Beta;
 import com.powsybl.iidm.network.Identifiable;
 import lombok.*;
+import org.apache.commons.lang3.Strings;
 import org.gridsuite.filter.utils.expertfilter.ExpertFilterUtils;
 import org.gridsuite.filter.utils.expertfilter.FieldType;
 import org.gridsuite.filter.utils.expertfilter.OperatorType;
-import org.gridsuite.filter.wip.expert.data.DataType;
+import org.gridsuite.filter.wip.data.DataType;
 
+import java.util.Collections;
 import java.util.Objects;
-import java.util.Optional;
+import java.util.Set;
 
 /**
  * @author Kamil MARUT {@literal <kamil.marut at rte-france.com>}
  */
 @Beta
 @Data
+@EqualsAndHashCode
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-@EqualsAndHashCode(callSuper = true)
-public final class BooleanExpertRule extends AbstractExpertRule {
+public final class StringExpertRule implements ExpertRule {
 
     private FieldType fieldType;
     private OperatorType operatorType;
-    private Boolean referenceValue;
+    private Set<String> referenceValues;
+    private String referenceValue;
 
     @Builder
-    public BooleanExpertRule(FieldType fieldType, OperatorType operatorType, Boolean referenceValue) {
+    public StringExpertRule(FieldType fieldType, OperatorType operatorType, String referenceValue, Set<String> referenceValues) {
         this.fieldType = Objects.requireNonNull(fieldType);
         this.operatorType = Objects.requireNonNull(operatorType);
-        this.referenceValue = referenceValue;
+        this.referenceValue = referenceValue != null ? referenceValue : "";
+        this.referenceValues = referenceValues != null ? referenceValues : Collections.emptySet();
     }
 
     @Override
     public boolean evaluateRule(Identifiable<?> identifiable) {
         String fieldValue = ExpertFilterUtils.getFieldValue(fieldType, null, identifiable);
-        if (fieldValue == null) {
-            return OperatorType.NOT_EXISTS.equals(operatorType);
+        if (fieldValue == null || fieldValue.isEmpty()) {
+            return operatorType.equals(OperatorType.NOT_EXISTS);
         }
 
-        boolean parsedFieldValue = Boolean.parseBoolean(fieldValue);
-        boolean parsedReferenceValue = Optional.ofNullable(referenceValue).orElse(false);
         return switch (operatorType) {
-            case EQUALS -> parsedReferenceValue == parsedFieldValue;
-            case NOT_EQUALS -> parsedReferenceValue != parsedFieldValue;
+            case IS -> Strings.CI.equals(fieldValue, referenceValue);
+            case CONTAINS -> Strings.CI.contains(fieldValue, referenceValue);
+            case BEGINS_WITH -> Strings.CI.startsWith(fieldValue, referenceValue);
+            case ENDS_WITH -> Strings.CI.endsWith(fieldValue, referenceValue);
             case EXISTS -> true;
             case NOT_EXISTS -> false;
+            case IN -> referenceValues.stream().anyMatch(fieldValue::equalsIgnoreCase);
+            case NOT_IN -> referenceValues.stream().noneMatch(fieldValue::equalsIgnoreCase);
             default -> throw unsupportedOperatorException();
         };
     }
@@ -61,11 +67,6 @@ public final class BooleanExpertRule extends AbstractExpertRule {
     @Override
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
     public DataType getDataType() {
-        return DataType.BOOLEAN;
-    }
-
-    @Override
-    public OperatorType getOperatorType() {
-        return operatorType;
+        return DataType.STRING;
     }
 }
